refactor(bongocat.c): move bongo cat OLED code to second file
